move fixtures; linting
