+++ core/cat/looking_glass/prompts.py
@@ -1,7 +1,3 @@
-import json
-import random
-from typing import Union, Dict
-
 
 TOOL_PROMPT = """Create a JSON action to complete the action sequence, with the correct "action" and "action_input" to help the Human.
 You can use one of these actions:
