+++ core/tests/agents/test_agent_form.py
@@ -1,7 +1,6 @@
 
 import pytest
 
-from tests.agents.agent_fixtures import main_agent, stray
 
 @pytest.mark.asyncio
 async def test_execute_form_agent(main_agent, stray):
