+++ core/tests/agents/test_agent_memory.py
@@ -1,9 +1,7 @@
-import pytest
 import time
 
 from langchain.docstore.document import Document
 
-from tests.agents.agent_fixtures import main_agent, stray
 
 
 def test_format_agent_input_on_empty_memory(main_agent, stray):
