+++ core/tests/agents/test_output_parser.py
@@ -1,4 +1,3 @@
-import pytest
 
 from langchain.schema import AgentAction, AgentFinish
