+++ core/tests/mocks/mock_plugin/mock_llm.py
@@ -1,5 +1,5 @@
 
-from typing import List, Type, Dict
+from typing import List, Type
 from pydantic import ConfigDict
 
 from langchain_community.llms.fake import FakeListLLM
